Adds menu item for reapplying the store mesh z offset
A menu item "Old Z Position" is added to the manual bed leveling
menu so it is possible to reuse the old value when relevelling the
bed.
It can for instance be used to quickly go trough the values to correct
one or two values which were set incorrectly due to pressing next point
by mistake.